# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Group location tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface Group {
  id: string;
  name: string;
  inviteCode: string;
  createdAt: string;
  memberCount: number;
}

export interface GroupWithToken {
  group: Group;
  memberId: string;
  token: string;
}

export interface GroupInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  creatorName: string;
}

export interface JoinGroupInput {
  inviteCode: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  memberName: string;
}

export interface Member {
  id: string;
  name: string;
  groupId: string;
  isLocationSharing: boolean;
  joinedAt: string;
}

export interface MemberLocation {
  memberId: string;
  memberName: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  accuracy?: number | null;
  updatedAt: string;
  isSharing: boolean;
}

export interface LocationInput {
  latitude: number;
  longitude: number;
  /** @nullable */
  accuracy?: number | null;
}

export interface LocationSharingInput {
  isSharing: boolean;
}

export interface MeetingPoint {
  id: string;
  groupId: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  label?: string | null;
  setByName: string;
  setAt: string;
}

export interface MeetingPointInput {
  latitude: number;
  longitude: number;
  /** @nullable */
  label?: string | null;
}

export type MessageType = typeof MessageType[keyof typeof MessageType];


export const MessageType = {
  chat: 'chat',
  system: 'system',
  sos: 'sos',
  meeting_point: 'meeting_point',
} as const;

export interface Message {
  id: string;
  groupId: string;
  /** @nullable */
  memberId?: string | null;
  memberName: string;
  content: string;
  type: MessageType;
  createdAt: string;
}

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}