# lib/api-zod/src/generated/types/groupWithToken.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Group location tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { Group } from './group';

export interface GroupWithToken {
  group: Group;
  memberId: string;
  token: string;
}