# lib/api-zod/src/generated/types/joinGroupInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Group location tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface JoinGroupInput {
  inviteCode: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  memberName: string;
}