# lib/api-zod/src/generated/types/locationSharingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Group location tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface LocationSharingInput {
  isSharing: boolean;
}