# lib/api-zod/src/generated/types/meetingPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Group location tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface MeetingPoint {
  id: string;
  groupId: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  label?: string | null;
  setByName: string;
  setAt: string;
}