# lib/api-zod/src/generated/types/meetingPointInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Group location tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface MeetingPointInput {
  latitude: number;
  longitude: number;
  /** @nullable */
  label?: string | null;
}