# lib/api-zod/src/generated/types/memberLocation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Group location tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface MemberLocation {
  memberId: string;
  memberName: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  accuracy?: number | null;
  updatedAt: string;
  isSharing: boolean;
}