# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Group location tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageType } from './messageType';

export interface Message {
  id: string;
  groupId: string;
  /** @nullable */
  memberId?: string | null;
  memberName: string;
  content: string;
  type: MessageType;
  createdAt: string;
}